Remove more visual studio compiler warnings.

+++ src-library/Binasc.cpp
@@ -1393,9 +1393,9 @@ int Binasc::processDecimalWord(std::ostream& out, const std::string& word,
 				return 0;
 			}
 			ulong tempLong = (ulong)atoi(&word[quoteIndex + 1]);
-			uchar byte1 = (tempLong & 0x00ff0000) >> 16;
-			uchar byte2 = (tempLong & 0x0000ff00) >>  8;
-			uchar byte3 = (tempLong & 0x000000ff);
+			uchar byte1 = (uchar)((tempLong & 0x00ff0000) >> 16);
+			uchar byte2 = (uchar)((tempLong & 0x0000ff00) >> 8);
+			uchar byte3 = (uchar)((tempLong & 0x000000ff));
 			if (endianIndex == -1) {
 				out << byte1;
 				out << byte2;
